Move tests to the correct file
