+++ tests/components/openai_conversation/test_init.py
@@ -1,6 +1,6 @@
 """Tests for the OpenAI integration."""
 
-from unittest.mock import AsyncMock, patch
+from unittest.mock import patch
 
 from httpx import Response
 from openai import (
@@ -9,263 +9,15 @@ from openai import (
     BadRequestError,
     RateLimitError,
 )
-from openai.types.chat.chat_completion import ChatCompletion, Choice
-from openai.types.chat.chat_completion_message import ChatCompletionMessage
-from openai.types.chat.chat_completion_message_tool_call import (
-    ChatCompletionMessageToolCall,
-    Function,
-)
-from openai.types.completion_usage import CompletionUsage
 from openai.types.image import Image
 from openai.types.images_response import ImagesResponse
 import pytest
-import voluptuous as vol
 
-from homeassistant.components import conversation
-from homeassistant.core import Context, HomeAssistant
+from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers import intent, llm
 from homeassistant.setup import async_setup_component
 
 from tests.common import MockConfigEntry
-
-
-@patch("homeassistant.components.openai_conversation.conversation.llm.async_get_tools")
-@patch("homeassistant.components.openai_conversation.conversation.llm.async_call_tool")
-async def test_function_call(
-    mock_call_tool,
-    mock_get_tools,
-    hass: HomeAssistant,
-    mock_config_entry: MockConfigEntry,
-    mock_init_component,
-) -> None:
-    """Test function call from the assistant."""
-    entry = MockConfigEntry(title=None)
-    entry.add_to_hass(hass)
-
-    mock_call_tool.return_value = "Test response"
-
-    mock_tool = llm.Tool()
-    mock_tool.name = "test_tool"
-    mock_tool.description = "Test function"
-    mock_tool.parameters = vol.Schema(
-        {vol.Optional("param1", description="Test parameters"): str}
-    )
-
-    mock_get_tools.return_value = [mock_tool]
-
-    def completion_result(*args, messages, **kwargs):
-        for message in messages:
-            role = message["role"] if isinstance(message, dict) else message.role
-            if role == "tool":
-                return ChatCompletion(
-                    id="chatcmpl-1234567890ZYXWVUTSRQPONMLKJIH",
-                    choices=[
-                        Choice(
-                            finish_reason="stop",
-                            index=0,
-                            message=ChatCompletionMessage(
-                                content="I have successfully called the function",
-                                role="assistant",
-                                function_call=None,
-                                tool_calls=None,
-                            ),
-                        )
-                    ],
-                    created=1700000000,
-                    model="gpt-4-1106-preview",
-                    object="chat.completion",
-                    system_fingerprint=None,
-                    usage=CompletionUsage(
-                        completion_tokens=9, prompt_tokens=8, total_tokens=17
-                    ),
-                )
-
-        return ChatCompletion(
-            id="chatcmpl-1234567890ABCDEFGHIJKLMNOPQRS",
-            choices=[
-                Choice(
-                    finish_reason="tool_calls",
-                    index=0,
-                    message=ChatCompletionMessage(
-                        content=None,
-                        role="assistant",
-                        function_call=None,
-                        tool_calls=[
-                            ChatCompletionMessageToolCall(
-                                id="call_AbCdEfGhIjKlMnOpQrStUvWx",
-                                function=Function(
-                                    arguments='{"param1":"test_value"}',
-                                    name="test_tool",
-                                ),
-                                type="function",
-                            )
-                        ],
-                    ),
-                )
-            ],
-            created=1700000000,
-            model="gpt-4-1106-preview",
-            object="chat.completion",
-            system_fingerprint=None,
-            usage=CompletionUsage(
-                completion_tokens=9, prompt_tokens=8, total_tokens=17
-            ),
-        )
-
-    context = Context()
-
-    with patch(
-        "openai.resources.chat.completions.AsyncCompletions.create",
-        new_callable=AsyncMock,
-        side_effect=completion_result,
-    ) as mock_create:
-        result = await conversation.async_converse(
-            hass,
-            "Please call the test function",
-            None,
-            context,
-            agent_id=mock_config_entry.entry_id,
-        )
-
-    assert result.response.response_type == intent.IntentResponseType.ACTION_DONE
-    assert mock_create.mock_calls[1][2]["messages"][3] == {
-        "role": "tool",
-        "tool_call_id": "call_AbCdEfGhIjKlMnOpQrStUvWx",
-        "name": "test_tool",
-        "content": '"Test response"',
-    }
-    mock_call_tool.assert_awaited_once_with(
-        hass,
-        llm.ToolInput(
-            tool_name="test_tool",
-            tool_args={"param1": "test_value"},
-            platform="openai_conversation",
-            context=context,
-            user_prompt="Please call the test function",
-            language="en",
-            assistant="conversation",
-        ),
-    )
-
-
-@patch("homeassistant.components.openai_conversation.conversation.llm.async_get_tools")
-@patch("homeassistant.components.openai_conversation.conversation.llm.async_call_tool")
-async def test_function_exception(
-    mock_call_tool,
-    mock_get_tools,
-    hass: HomeAssistant,
-    mock_config_entry: MockConfigEntry,
-    mock_init_component,
-) -> None:
-    """Test function call with exception."""
-    entry = MockConfigEntry(title=None)
-    entry.add_to_hass(hass)
-
-    mock_call_tool.side_effect = RuntimeError("Test tool exception")
-
-    mock_tool = llm.Tool()
-    mock_tool.name = "test_tool"
-    mock_tool.description = "Test function"
-    mock_tool.parameters = vol.Schema(
-        {vol.Optional("param1", description="Test parameters"): str}
-    )
-
-    mock_get_tools.return_value = [mock_tool]
-
-    def completion_result(*args, messages, **kwargs):
-        for message in messages:
-            role = message["role"] if isinstance(message, dict) else message.role
-            if role == "tool":
-                return ChatCompletion(
-                    id="chatcmpl-1234567890ZYXWVUTSRQPONMLKJIH",
-                    choices=[
-                        Choice(
-                            finish_reason="stop",
-                            index=0,
-                            message=ChatCompletionMessage(
-                                content="There was an error calling the function",
-                                role="assistant",
-                                function_call=None,
-                                tool_calls=None,
-                            ),
-                        )
-                    ],
-                    created=1700000000,
-                    model="gpt-4-1106-preview",
-                    object="chat.completion",
-                    system_fingerprint=None,
-                    usage=CompletionUsage(
-                        completion_tokens=9, prompt_tokens=8, total_tokens=17
-                    ),
-                )
-
-        return ChatCompletion(
-            id="chatcmpl-1234567890ABCDEFGHIJKLMNOPQRS",
-            choices=[
-                Choice(
-                    finish_reason="tool_calls",
-                    index=0,
-                    message=ChatCompletionMessage(
-                        content=None,
-                        role="assistant",
-                        function_call=None,
-                        tool_calls=[
-                            ChatCompletionMessageToolCall(
-                                id="call_AbCdEfGhIjKlMnOpQrStUvWx",
-                                function=Function(
-                                    arguments='{"param1":"test_value"}',
-                                    name="test_tool",
-                                ),
-                                type="function",
-                            )
-                        ],
-                    ),
-                )
-            ],
-            created=1700000000,
-            model="gpt-4-1106-preview",
-            object="chat.completion",
-            system_fingerprint=None,
-            usage=CompletionUsage(
-                completion_tokens=9, prompt_tokens=8, total_tokens=17
-            ),
-        )
-
-    context = Context()
-
-    with patch(
-        "openai.resources.chat.completions.AsyncCompletions.create",
-        new_callable=AsyncMock,
-        side_effect=completion_result,
-    ) as mock_create:
-        result = await conversation.async_converse(
-            hass,
-            "Please call the test function",
-            None,
-            context,
-            agent_id=mock_config_entry.entry_id,
-        )
-
-    assert result.response.response_type == intent.IntentResponseType.ACTION_DONE
-    assert mock_create.mock_calls[1][2]["messages"][3] == {
-        "role": "tool",
-        "tool_call_id": "call_AbCdEfGhIjKlMnOpQrStUvWx",
-        "name": "test_tool",
-        "content": '{"error": "RuntimeError", "error_text": "Test tool exception"}',
-    }
-    mock_call_tool.assert_awaited_once_with(
-        hass,
-        llm.ToolInput(
-            tool_name="test_tool",
-            tool_args={"param1": "test_value"},
-            platform="openai_conversation",
-            context=context,
-            user_prompt="Please call the test function",
-            language="en",
-            assistant="conversation",
-        ),
-    )
 
 
 @pytest.mark.parametrize(
